api/requirements.txt
# FastAPI + MongoDB Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# MongoDB Async Driver
motor==3.3.2
pymongo==4.6.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Additional Dependencies
python-multipart==0.0.6

# Machine Learning
pandas==2.1.4
joblib==1.3.2
scikit-learn==1.3.2  # For model compatibility

# Development Tools (optional)
# pytest==7.4.3
# httpx==0.26.0
